Add support for multiple suffixes in clang-tidy
By default, clang-tidy only support one possible suffix.

I ran into a case where I needed clang-tidy to be able to check between different suffixes.
In that case, a class might end with Suffix1, Suffix2, ...
This patch is a quick and dirty way to support this feature using '|' as separator for the
suffixes in the options.

Known limitation: if 'AClass' does not respect the convention, clang-tidy will suggest this fix:
AClassSuffix1|Suffix2
It would be better to display the following suggestion:
AClass(Suffix1|Suffix2)

This is also a problem if we use the '-fix' option.
A possibility would be to use the first suffix in the list.